raspberry/requirements.txt
pyserial==3.5.0
gpiozero==2.0.1